test(e2e): add snapshotStory utility
